+++ src/main/java/com/example/demo/controller/ScrapController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.demo.service.Scraper;
+import com.example.demo.scraper.Scraper;
 
 @RestController
 @RequestMapping("/scrap")
