+++ src/main/java/com/example/demo/scraper/Scraper.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.scraper;
 
 public interface Scraper {
 	public void scrap();
